+++ ReleaseNotes.md
@@ -2,4 +2,7 @@
 * Initial release
 
 ### New in 0.0.1 (Released 2015/10/07)
-* Comments and adding dependencies to nuget
+* Comments and adding dependencies to NuGet
+
+### New in 0.0.2 (Released 2015/11/16)
+* Store Redis ConnectionMultiplexer per config for single use